Standartizing broker and OTEL processors
