+++ ark/internal/telemetry/routing/processor.go
@@ -3,18 +3,10 @@ package routing
 import (
 	"context"
 	"strings"
-	"sync"
 
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp"
 	"go.opentelemetry.io/otel/sdk/trace"
-
-	"mckinsey.com/ark/internal/telemetry"
 )
-
-type RoutingSpanProcessor struct {
-	endpoints map[string]*ExporterConfig
-	mu        sync.RWMutex
-}
 
 type ExporterConfig struct {
 	Namespace string
@@ -22,10 +14,8 @@ type ExporterConfig struct {
 	Processor trace.SpanProcessor
 }
 
-func NewRoutingSpanProcessor(ctx context.Context, endpoints []BrokerEndpoint) (*RoutingSpanProcessor, error) {
-	rsp := &RoutingSpanProcessor{
-		endpoints: make(map[string]*ExporterConfig),
-	}
+func NewRoutingSpanProcessor(ctx context.Context, endpoints []BrokerEndpoint) (*NamespaceRoutingProcessor, error) {
+	configs := make(map[string]*ExporterConfig)
 
 	for _, endpoint := range endpoints {
 		otlpEndpoint := endpoint.Endpoint + "/v1/traces"
@@ -37,7 +27,7 @@ func NewRoutingSpanProcessor(ctx context.Context, endpoints []BrokerEndpoint) (*
 			continue
 		}
 
-		rsp.endpoints[endpoint.Namespace] = &ExporterConfig{
+		configs[endpoint.Namespace] = &ExporterConfig{
 			Namespace: endpoint.Namespace,
 			Exporter:  exporter,
 			Processor: trace.NewBatchSpanProcessor(exporter),
@@ -46,7 +36,7 @@ func NewRoutingSpanProcessor(ctx context.Context, endpoints []BrokerEndpoint) (*
 		log.Info("created broker exporter", "namespace", endpoint.Namespace)
 	}
 
-	return rsp, nil
+	return NewNamespaceRoutingProcessor(configs), nil
 }
 
 func createExporter(ctx context.Context, endpoint string) (trace.SpanExporter, error) {
@@ -81,29 +71,6 @@ func extractPath(url string) string {
 	return "/"
 }
 
-func (r *RoutingSpanProcessor) OnStart(parent context.Context, s trace.ReadWriteSpan) {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	for _, config := range r.endpoints {
-		config.Processor.OnStart(parent, s)
-	}
-}
-
-func (r *RoutingSpanProcessor) OnEnd(s trace.ReadOnlySpan) {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	queryNamespace := getStringAttribute(s, telemetry.AttrQueryNamespace)
-	if queryNamespace == "" {
-		return
-	}
-
-	if config, ok := r.endpoints[queryNamespace]; ok {
-		config.Processor.OnEnd(s)
-	}
-}
-
 func getStringAttribute(span trace.ReadOnlySpan, key string) string {
 	for _, attr := range span.Attributes() {
 		if string(attr.Key) == key {
@@ -111,28 +78,4 @@ func getStringAttribute(span trace.ReadOnlySpan, key string) string {
 		}
 	}
 	return ""
-}
-
-func (r *RoutingSpanProcessor) Shutdown(ctx context.Context) error {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-
-	for _, config := range r.endpoints {
-		if err := config.Processor.Shutdown(ctx); err != nil {
-			log.Error(err, "failed to shutdown processor", "namespace", config.Namespace)
-		}
-	}
-	return nil
-}
-
-func (r *RoutingSpanProcessor) ForceFlush(ctx context.Context) error {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	for _, config := range r.endpoints {
-		if err := config.Processor.ForceFlush(ctx); err != nil {
-			log.Error(err, "failed to flush processor", "namespace", config.Namespace)
-		}
-	}
-	return nil
 }
